Shuffle files to the packages that correspond to library package/groupId
