+++ src/main/java/com/simprints/biometrics/simface/Constants.kt
@@ -1,0 +1,7 @@
+package com.simprints.biometrics.simface
+
+internal object Constants {
+    const val IMAGE_SIZE = 112
+    const val OUTPUT_EMBEDDING_SIZE = 512
+    const val DEFAULT_TEMPLATE_VERSION = "SIM_FACE_BASE_1"
+}
