+++ src/main/java/com/simprints/biometrics/simface/data/Point2D.kt
@@ -1,0 +1,6 @@
+package com.simprints.biometrics.simface.data
+
+data class Point2D(
+    val x: Float,
+    val y: Float,
+)
