+++ src/main/java/com/simprints/biometrics/simface/embedding/EmbeddingProcessor.kt
@@ -1,0 +1,7 @@
+package com.simprints.biometrics.simface.embedding
+
+import android.graphics.Bitmap
+
+internal interface EmbeddingProcessor {
+    fun getEmbedding(bitmap: Bitmap): ByteArray
+}
