--- src/main/java/com/simprints/simface/data/Point2D.kt
@@ -1,6 +1,0 @@
-package com.simprints.simface.data
-
-data class Point2D(
-    val x: Float,
-    val y: Float,
-)
