--- src/main/java/com/simprints/simface/matcher/MatchProcessor.kt
@@ -1,13 +1,0 @@
-package com.simprints.simface.matcher
-
-internal interface MatchProcessor {
-    fun verificationScore(
-        probe: ByteArray,
-        matchAgainst: ByteArray,
-    ): Double
-
-    fun identificationScore(
-        probe: ByteArray,
-        matchReferences: List<ByteArray>,
-    ): List<Pair<ByteArray, Double>>
-}
